data/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
faiss-cpu==1.7.4
numpy==1.21.2
pandas==1.3.3
sentence-transformers==2.2.2
torch==1.9.1
transformers==4.11.3
nest-asyncio==1.5.1
pydantic==1.8.2
sqlite3==2.6.0 